+++ src/mint.rs
@@ -1,7 +1,7 @@
 use anyhow::{anyhow, Result};
 use glob::glob;
 use log::{error, info};
-use metaplex_token_metadata::instruction::{
+use mpl_token_metadata::instruction::{
     create_master_edition, create_metadata_accounts, update_metadata_accounts,
 };
 use rayon::prelude::*;